python/requirements.txt
# Chess Analysis Dependencies
python-chess==1.999
stockfish==3.28.0
supabase>=2.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pandas>=2.1.3
numpy>=1.25.2
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
pytest>=7.4.0
flake8>=6.0.0
stripe>=7.0.0
psutil>=5.9.0
anthropic>=0.18.0
